modularize inventory and network forms
