+++ modules/inventory/__init__.py
@@ -1,4 +1,6 @@
 """Inventory module exports."""
 
 # Router is defined in modules.inventory.routes
-__all__ = []
+from . import forms, utils
+
+__all__ = ["forms", "utils"]
