+++ modules/network/__init__.py
@@ -1,4 +1,6 @@
 """Network module exports."""
 
 # Router is defined in modules.network.routes
-__all__ = []
+from . import forms
+
+__all__ = ["forms"]
